Addressing PR feedback
thank you copilot for the lovely feedback
