+++ src/app/home-page/ask-morgan-form/ask-morgan-form.html
@@ -1,11 +1,13 @@
 <div class="ask-morgan-container" [class.success-state]="isSuccess">
-  <div *ngIf="!isSuccess">
+  @if (!isSuccess) {
 
-    <div *ngIf="errorMessage" class="error-message" aria-live="assertive">
-      {{ errorMessage }}
-    </div>
+    @if (errorMessage) {
+      <div class="error-message" aria-live="assertive">
+        {{ errorMessage }}
+      </div>
+    }
 
-    <form (ngSubmit)="submitForm()" #askForm="ngForm">
+    <form (ngSubmit)="submitForm()">
       <div class="honey-pot" aria-hidden="true">
         <label for="username_honey">Leave this field blank if you are human:</label>
         <input type="text" id="username_honey" name="username_honey" [(ngModel)]="username_honey" tabindex="-1" autocomplete="off">
@@ -36,9 +38,10 @@
         {{ isSubmitting ? 'Sending...' : 'Submit' }}
       </button>
     </form>
-  </div>
+  }
 
-  <div *ngIf="isSuccess" class="success-feedback" aria-live="polite">
+  @if (isSuccess) {
+    <div class="success-feedback" aria-live="polite">
     <div class="success-icon" aria-hidden="true">
       <svg viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round">
         <path d="M22 11.08V12a10 10 0 1 1-5.93-9.14"></path>
@@ -48,5 +51,6 @@
     <h4>Message Sent</h4>
     <p>{{ successMessage }}</p>
     <button type="button" class="reset-btn" (click)="resetForm()">Ask Another Question</button>
-  </div>
+    </div>
+  }
 </div>
