+++ src/app/home-page/ask-morgan-form/ask-morgan-form.scss
@@ -1,4 +1,4 @@
-@use '../../styles' as *;
+@use '../../../styles' as *;
 
 .ask-morgan-container {
   box-sizing: border-box;
