+++ src/app/home-page/home-page.component.html
@@ -52,7 +52,7 @@
 
       <section>
         <h2>Contact</h2>
-        <p>Feel free to reach out via email or at any of the other platforms linked here. Or if you'd prefer an anonymous submission, you can use the following:</p>
+        <p>Feel free to reach out via email at <a href="mailto:website@morganvanv.com" class="pill">website&#64;morganvanv.com</a> or at any of the other platforms linked here. Or if you'd prefer an anonymous submission, you can use the following:</p>
         <app-ask-morgan-form></app-ask-morgan-form>
       </section>
